+++ packages/eyes-playwright-universal/.mocharc.cvg.js
@@ -1,6 +1,6 @@
 const {mochaGrep} = require('@applitools/test-utils')
 
-const tags = ['chrome', 'chromium', 'firefox', 'webkit', 'safari']
+const tags = ['chrome', 'chromium', 'firefox', 'webkit', 'safari', 'all-cookies']
 
 module.exports = {
   spec: [