+++ packages/eyes-playwright-universal/src/universal-client.ts
@@ -25,14 +25,14 @@ export class UniversalClient implements types.Core<Driver, Element, Selector> {
     this._refer = new Refer((value: any): value is Driver | Context | Element => {
       return spec.isDriver(value) || spec.isContext(value) || spec.isElement(value)
     })
-    this._server = spawn('node', ['./node_modules/@applitools/eyes-universal/dist/cli.js', '--port=2107'], {
+    this._server = spawn('node', ['./node_modules/@applitools/eyes-universal/dist/cli.js'], {
       detached: true,
       stdio: ['ignore', 'pipe', 'ignore'],
     })
 
     // specific to JS: we are able to listen to stdout for the first line, then we know the server is up, and we even can get its port in case it wasn't passed
     this._server.stdout.once('data', data => {
-      this._server.stdout.destroy()
+      ;(this._server.stdout as any).unref()
       const [port] = String(data).split('\n', 1)
       this._socket.connect(`http://localhost:${port}/eyes`)
       this._socket.emit('Core.makeSDK', {