+++ packages/driver/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- made `setViewportSize` more reliable in worst case scenario and faster in best case scenario
 
 ## 1.3.5 - 2021/11/23