+++ packages/eyes-selenium-universal/CHANGELOG.md
@@ -6,6 +6,10 @@
 - updated to @applitools/eyes-api@1.1.2 (from 1.1.1)
 - updated to @applitools/types@1.0.14 (from 1.0.13)
 - updated to @applitools/utils@1.2.3 (from 1.2.0)
+- updated to @applitools/eyes-api@1.1.6 (from 1.1.2)
+- updated to @applitools/eyes-universal@0.2.0 (from 0.1.4)
+- updated to @applitools/types@1.0.21 (from 1.0.14)
+- updated to @applitools/utils@1.2.4 (from 1.2.3)
 
 ## 1.2.2 - 2020/10/15