+++ packages/logger/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- change log file name in rolling file handler to not contain windows forbidden characters.
 
 ## 1.0.6 - 2021/11/23